refactor: redireciona automaticamente se o usuário já estiver logado e tentar acessar o login/register
