+++ client/src/app/router.tsx
@@ -1,61 +1,7 @@
-import { type QueryClient, useQueryClient } from '@tanstack/react-query';
+import { useQueryClient } from '@tanstack/react-query';
 import { useMemo } from 'react';
-import { createBrowserRouter } from 'react-router';
 import { RouterProvider } from 'react-router/dom';
-import { paths } from '@/config/paths';
-import { ProtectedRoute } from '@/lib/auth';
-import { default as AppRoot, ErrorBoundary as AppRootErrorBoundary } from './routes/app/root';
-
-const convert = (queryClient: QueryClient) => (m: any) => {
-	const { clientLoader, clientAction, default: Component, ...rest } = m;
-	return {
-		...rest,
-		loader: clientLoader?.(queryClient),
-		action: clientAction?.(queryClient),
-		Component,
-	};
-};
-
-export const createAppRouter = (queryClient: QueryClient) =>
-	createBrowserRouter([
-		{
-			path: '/',
-			HydrateFallback: () => null,
-			children: [
-				{
-					path: paths.home.path,
-					lazy: () => import('./routes/landing').then(convert(queryClient)),
-				},
-				{
-					path: paths.auth.register.path,
-					lazy: () => import('./routes/auth/register').then(convert(queryClient)),
-				},
-				{
-					path: paths.auth.login.path,
-					lazy: () => import('./routes/auth/login').then(convert(queryClient)),
-				},
-				{
-					path: paths.app.root.path,
-					element: (
-						<ProtectedRoute>
-							<AppRoot />
-						</ProtectedRoute>
-					),
-					ErrorBoundary: AppRootErrorBoundary,
-					children: [
-						{
-							path: paths.app.dashboard.path,
-							lazy: () => import('./routes/app/dashboard').then(convert(queryClient)),
-						},
-					],
-				},
-				{
-					path: '*',
-					lazy: () => import('./routes/not-found').then(convert(queryClient)),
-				},
-			],
-		},
-	]);
+import { createAppRouter } from './create-app-router';
 
 export const AppRouter = () => {
 	const queryClient = useQueryClient();
