--- client/src/app/routes/app/root.tsx
@@ -1,13 +1,0 @@
-import { Outlet } from 'react-router';
-
-export function ErrorBoundary() {
-	return <div>Something went wrong!</div>;
-}
-
-export default function AppRoot() {
-	return (
-		<div>
-			<Outlet />
-		</div>
-	);
-}
